# Rewrite `vec_equal()`